Added french translation
I translated as much as I could in French (France, Canada).
French translation of certain options can be incredibly long with too many words, so I had to simplify certain menu (like removing the repeating word "shadow" in sub menu for shadows, or find a way to make the colour schemes easy to understand with the use of colours).
I also added comments for every option with a description of what this option does, and what impact it may have on performances when choosing an option instead of another.
I hope I did things well; I love this shader, so I wanted to help a bit. I am new on GitHub for this occasion.